+++ src/components/InputIncomeExpense.vue
@@ -74,13 +74,14 @@ const categories = ref([]);
 const accountStore = useAccountListStore();
 const calendarAccountStore = useCalendarAccountStore(); //calendar
 const { accountList, fetchAccountList, addAccount } = accountStore;
-const { fetchSummaryList } = calendarAccountStore;
+const { fetchSummaryList, fetchDailyAccountList } = calendarAccountStore;
 
 // 추가 버튼 눌러서 추가되면 입력창 리셋
-const successCallback = () => {
+const successCallback = (date = null) => {
   resetForm();
   fetchAccountList();
   fetchSummaryList();
+  if(date != null) fetchDailyAccountList(date);
 };
 
 const handleAddAccount = async () => {
